+++ src/migrations/Migrators/BlogChannelMigrator.php
@@ -13,11 +13,15 @@ class BlogChannelMigrator extends Migrator
 {
     public static function add(): bool
     {
+        \Craft::$app->cache->delete(Handles::CHANNEL);
+
         return (new SectionService())->add('Blog', Handles::CHANNEL, Section::TYPE_CHANNEL, '/{slug}', 'post/_entry');
     }
 
     public static function remove(): bool
     {
+        \Craft::$app->cache->delete(Handles::CHANNEL);
+
         return (new SectionService())->remove(Handles::CHANNEL);
     }
 }