+++ src/twigextensions/BlogTwigExtension.php
@@ -6,6 +6,8 @@ use craft\elements\Category;
 use craft\elements\Entry;
 use craft\elements\Tag;
 use matfish\Blogify\Handles;
+use matfish\Blogify\services\IdsService;
+use matfish\Blogify\services\PostViewsService;
 use Twig\Extension\AbstractExtension;
 use Twig\Extension\GlobalsInterface;
 use Twig\TwigFunction;
@@ -15,11 +17,13 @@ class BlogTwigExtension extends AbstractExtension implements GlobalsInterface
 
     public function getGlobals()
     {
+        $idsService = new IdsService();
+
         return [
             'blogify' => [
-                'categories' => Category::find()->groupId($this->getCategoriesGroupId()),
+                'categories' => Category::find()->groupId($idsService->categoriesGroupId()),
                 'usedCategories' => $this->getUsedCategories(),
-                'tags' => Tag::find()->groupId($this->getTagsGroupId())
+                'tags' => Tag::find()->groupId($idsService->tagsGroupId())
             ]
         ];
     }
@@ -28,7 +32,8 @@ class BlogTwigExtension extends AbstractExtension implements GlobalsInterface
     {
         return [
             new TwigFunction('blogifyRecentPosts', [$this, 'getRecentPosts']),
-            new TwigFunction('blogifySearch', [$this, 'search'])
+            new TwigFunction('blogifySearch', [$this, 'search']),
+            new TwigFunction('blogifyPopularPosts', [$this, 'getPopularPosts'])
         ];
     }
 
@@ -39,6 +44,21 @@ class BlogTwigExtension extends AbstractExtension implements GlobalsInterface
             ->orderBy('postDate desc');
     }
 
+    public function getPopularPosts($excludeNoViews = false)
+    {
+        (new PostViewsService())->verifyEnabled();
+
+        $q = Entry::find()
+            ->section(Handles::CHANNEL)
+            ->orderBy(Handles::POST_VIEWS . ' desc');
+
+        if ($excludeNoViews) {
+            $q = $q->blogifyPostViews('>0');
+        }
+
+        return $q;
+    }
+
     public function search($query)
     {
         return Entry::find()
@@ -46,27 +66,12 @@ class BlogTwigExtension extends AbstractExtension implements GlobalsInterface
             ->search('"' . $query . '"');
     }
 
-
-    private function getTagsGroupId()
-    {
-        return \Craft::$app->cache->getOrSet(Handles::TAGS, function () {
-            return \Craft::$app->tags->getTagGroupByHandle(Handles::TAGS)->id;
-        }, 60 * 60 * 24 * 365);
-    }
-
-    private function getCategoriesGroupId()
-    {
-        return \Craft::$app->cache->getOrSet(Handles::CATEGORIES, function () {
-            return \Craft::$app->categories->getGroupByHandle(Handles::CATEGORIES)->id;
-        }, 60 * 60 * 24 * 365);
-    }
-
     private function getUsedCategories()
     {
         $entries = Entry::find()->section(Handles::CHANNEL);
 
         return Category::find()
-            ->groupId($this->getCategoriesGroupId())
+            ->groupId((new IdsService())->categoriesGroupId())
             ->relatedTo($entries);
     }
 }